+++ gui/index.html
@@ -99,7 +99,8 @@
 
 	<!-- basic imports (NEED TO BE LOCALIZED) -->
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-	<script src="http://pixijs.download/release/pixi.min.js"></script>
+
+	<!--TODO: localize d3-->
 	<script src="https://d3js.org/d3.v5.min.js"></script>
 
 	<!-- IMPORT CUSTOM BLOCKS -->
